+++ src/jormungandr/jormungandr/src/network/subscription.rs
@@ -2,7 +2,7 @@ use super::{buffer_sizes, convert::Decode, GlobalStateR};
 use crate::{
     blockcfg::Fragment,
     intercom::{self, BlockMsg, TopologyMsg, TransactionMsg},
-    network::bootstrap::local_addr_lookup,
+    network::retrieve_local_ip,
     settings::start::network::Configuration,
     topology::{Gossip, NodeId},
     utils::async_msg::{self, MessageBox},
@@ -380,7 +380,7 @@ impl FragmentProcessor {
             None => {
                 tracing::error!("node addr not present in config, reverting to local lookup");
                 FragmentOrigin::Network {
-                    addr: local_addr_lookup(),
+                    addr: retrieve_local_ip(),
                 }
             }
         };
